move Config classes into separate files
