+++ modules/kobalt-plugin-api/src/main/kotlin/com/beust/kobalt/api/IBuildConfig.kt
@@ -1,0 +1,11 @@
+package com.beust.kobalt.api
+
+interface IBuildConfig {
+    var buildConfig: BuildConfig?
+
+    fun buildConfig(init: BuildConfig.() -> Unit) {
+        buildConfig = BuildConfig().apply {
+            init()
+        }
+    }
+}
