+++ modules/kobalt-plugin-api/src/main/kotlin/com/beust/kobalt/api/Sources.kt
@@ -1,0 +1,11 @@
+package com.beust.kobalt.api
+
+import com.beust.kobalt.api.annotation.Directive
+import java.util.HashSet
+
+class Sources(val project: Project, val sources: HashSet<String>) {
+    @Directive
+    fun path(vararg paths: String) {
+        sources.addAll(paths)
+    }
+}
