+++ src/main/kotlin/com/beust/kobalt/plugin/apt/KaptConfig.kt
@@ -1,0 +1,5 @@
+package com.beust.kobalt.plugin.apt
+
+import com.beust.kobalt.api.annotation.Directive
+
+class KaptConfig(@Directive var outputDir: String = "generated/source/apt")
